.presets.json location
Before .presets.json was created and used from current working directory. This is not great because you might have different paths at different times.

Another option is to check the path of the file and create it in the same directory but I think that this OR in .config/pwc-cli/ should be better