checklist()
I had a lot of functions calling several of the checking functions before executing so I made a checklist function that takes user_input as a positional argument followed by strings (in the order they should be checked) of the tests that particular function needs to test.

checklist() then loops through the passed checks (based on a checklist_map) and runs the corresponding check functions. If any test fails it will return False and run the corresponding error-function. If all tests pass the function calling it will continue running.

Also cleaned up a bit.