Load preset bug + second chance
I have previously changed how the current values are presented. This had a side effect so that saved presets had unit names in the values. This meant that when loading settings they could not be set because the units are not in "valid settings" dictionary.

I am thinking about removing units from the results of get_current_settings() before saving them to the file but for now I just split on whitespace and send index 0 to change settings. I could also loop though the setting keys using is.digit() and save the result to a new variable.

Also offered user a second chance to enter a value and to change the value if it's not valid. Can fail once per check before returning to main screen.